Changed to uns short type data

+++ lib/mergeSort.cuh
@@ -7,7 +7,7 @@
 // Merges two subarrays of data[].
 // First subarray is data[l..m]
 // Second subarray is data[m+1..r]
-__host__ __device__ void merge(unsigned long *data, const unsigned long left, const unsigned long mid, const unsigned long right);
+__host__ __device__ void merge(unsigned short *data, const unsigned long long left, const unsigned long long mid, const unsigned long long right);
 
 /* Function that performs Merge Sort*/
-__host__ void merge_sort(unsigned long *data, const unsigned long left, const unsigned long right);
+__host__ void merge_sort(unsigned short *data, const unsigned long long left, const unsigned long long right);
